Update MedSAM finetuning.
- Add a validation set to measure performance.
A JSON output file is returned with the mean performance
of each validation batch. Also a curve is plotted
showing the mean validation performance for each epoch.
Validation performance is measured every 10 epochs
by default.
- Fix the extraction of bounding boxes from each
annotated slice. When the slice has more than one
2d object, one of them is randomly returned.
- The bounding box perturbation was reduced from
20 pixels to 5 pixels.
- Add the plotting of output segmentations for the
validation set. This was added as an optional flag
that is False by default.
- The default checkpoint model was changed from the
SAM model to the MedSAM model.